requirements.txt
pillow
opencv-python
mss
pyautogui
numpy